+++ apps-script/Trader.ts
@@ -4,15 +4,13 @@ import {Statistics} from "./Statistics";
 import {Config, IStore} from "./Store";
 import {ExchangeSymbol} from "./TradeResult";
 
-const PriceMemoMaxCapacity = 10;
-export type PriceMemo = [number, number, number]
-
 export class V2Trader {
   private readonly store: IStore;
   private readonly config: Config;
   private readonly exchange: IExchange;
   private readonly stats: Statistics;
   private readonly prices: { [p: string]: number };
+  private readonly afterAll: Array<() => void> = [];
 
   constructor(store: IStore, exchange: IExchange, stats: Statistics) {
     this.store = store;
@@ -22,72 +20,38 @@ export class V2Trader {
     this.prices = exchange.getPrices()
   }
 
+  /**
+   * After ticker check can be executed after all tickers are checked.
+   * It may contain any additional operations that need to be performed after all.
+   */
+  afterTickerCheck() {
+    while (this.afterAll.length > 0) {
+      this.afterAll.pop()();
+    }
+  }
+
   tickerCheck(tradeMemo: TradeMemo): void {
 
     if (tradeMemo.stateIs(TradeState.SELL)) {
-      this.sellAndClose(tradeMemo)
+      return this.sellAndClose(tradeMemo)
     }
 
     const symbol = tradeMemo.tradeResult.symbol;
-    const currentPrice = this.getPrice(symbol);
+    tradeMemo.pushPrice(this.getPrice(symbol))
 
-    tradeMemo.prices.push(currentPrice)
-    // remove old prices and keep only the last PriceMemoMaxCapacity
-    tradeMemo.prices.splice(0, tradeMemo.prices.length - PriceMemoMaxCapacity)
-    tradeMemo.maxObservedPrice = Math.max(tradeMemo.maxObservedPrice, ...tradeMemo.prices)
-
-    const priceGoesUp = this.priceGoesUp(tradeMemo.prices);
-
-    if (priceGoesUp) {
-      Log.info(`${symbol} price goes up`)
-    }
+    const priceGoesUp = tradeMemo.priceGoesUp()
+    priceGoesUp && Log.info(`${symbol} price goes up`)
 
     if (tradeMemo.stateIs(TradeState.BOUGHT)) {
-
-      if (currentPrice <= tradeMemo.stopLossPrice) {
-        const stopLimitCrossed = tradeMemo.prices[1] > tradeMemo.stopLossPrice;
-        if (stopLimitCrossed) {
-          Log.info(`${symbol}: crossed stop limit: price '${currentPrice}' <= '${tradeMemo.stopLossPrice}'`)
-        }
-        if (!tradeMemo.hodl && this.store.getConfig().SellAtStopLimit) {
-          tradeMemo.setState(TradeState.SELL)
-        }
-      }
-
-      const takeProfitPrice = tradeMemo.tradeResult.price * (1 + this.config.TakeProfit);
-      if (currentPrice >= takeProfitPrice) {
-        const takeProfitCrossed = tradeMemo.prices[1] < takeProfitPrice;
-        if (takeProfitCrossed) {
-          Log.alert(`${symbol} crossed take profit: price '${currentPrice}' >= '${takeProfitPrice}'`)
-        }
-        // We do not sell if price goes up, or we are in hodl mode, or if taking profit is disabled
-        if (!priceGoesUp && !tradeMemo.hodl && this.store.getConfig().SellAtTakeProfit) {
-          tradeMemo.setState(TradeState.SELL)
-        }
-      }
-
-      if (priceGoesUp) {
-        // Using previous price to calculate new stop limit
-        const newStopLimit = tradeMemo.prices[0] * (1 - this.config.LossLimit);
-        tradeMemo.stopLossPrice = tradeMemo.stopLossPrice < newStopLimit ? newStopLimit : tradeMemo.stopLossPrice
-      }
+      this.processBoughtState(tradeMemo);
+    } else if (tradeMemo.stateIs(TradeState.SOLD)) {
+      this.processSoldState(tradeMemo);
     }
 
-    if (tradeMemo.stateIs(TradeState.SOLD) && this.config.SwingTradeEnabled) {
-      // Swing trade enabled.
-      // Checking if price dropped below max observed price minus take profit percentage,
-      // and we can buy again
-      const priceDropped = currentPrice < tradeMemo.maxObservedPrice * (1 - this.config.TakeProfit);
-      if (priceDropped) {
-        tradeMemo.setState(TradeState.BUY)
-        Log.alert(`${symbol} will be bought again as price dropped sufficiently`)
-      } else {
-        Log.info(`${symbol} price has not dropped sufficiently, skipping swing trade`)
-      }
-    }
-
+    // save new pushed price and any intermediate state changes
     this.store.setTrade(tradeMemo)
 
+    // take action after processing
     if (tradeMemo.stateIs(TradeState.SELL)) {
       this.sellAndClose(tradeMemo)
     } else if (tradeMemo.stateIs(TradeState.BUY) && priceGoesUp) {
@@ -95,7 +59,51 @@ export class V2Trader {
     } else {
       Log.info(`${symbol}: waiting`)
     }
+  }
 
+  private processSoldState(tm: TradeMemo): void {
+    if (!this.config.SwingTradeEnabled) {
+      return
+    }
+    // Swing trade enabled.
+    // Checking if price dropped below max observed price minus profit limit percentage,
+    // and we can buy again
+    const symbol = tm.tradeResult.symbol;
+    const priceDropped = tm.currentPrice < tm.maxObservedPrice * (1 - this.config.ProfitLimit);
+    if (priceDropped) {
+      Log.alert(`${symbol} will be bought again as price dropped sufficiently`)
+      tm.setState(TradeState.BUY)
+    } else {
+      Log.info(`${symbol} price has not dropped sufficiently, skipping swing trade`)
+    }
+  }
+
+  private processBoughtState(tm: TradeMemo): void {
+    const symbol = tm.tradeResult.symbol;
+    const priceGoesUp = tm.priceGoesUp()
+
+    if (tm.profitLimitCrossedUp(this.config.ProfitLimit)) {
+      Log.alert(`${symbol} crossed profit limit`)
+    } else if (tm.lossLimitCrossedDown()) {
+      Log.alert(`${symbol}: crossed stop limit`)
+    }
+
+    if (tm.currentPrice < tm.stopLimitPrice) {
+      const canSell = !tm.hodl && this.store.getConfig().SellAtStopLimit;
+      canSell && tm.setState(TradeState.SELL)
+    }
+
+    const profitLimitPrice = tm.tradeResult.price * (1 + this.config.ProfitLimit);
+    if (tm.currentPrice > profitLimitPrice) {
+      const canSell = !tm.hodl && this.store.getConfig().SellAtProfitLimit;
+      canSell && !priceGoesUp && tm.setState(TradeState.SELL)
+    }
+
+    if (priceGoesUp) {
+      // Using previous price two measures back to calculate new stop limit
+      const newStopLimit = tm.prices[tm.prices.length - 3] * (1 - this.config.StopLimit);
+      tm.stopLimitPrice = tm.stopLimitPrice < newStopLimit ? newStopLimit : tm.stopLimitPrice
+    }
   }
 
   private getPrice(symbol: ExchangeSymbol): number {
@@ -111,21 +119,22 @@ export class V2Trader {
     if (tradeResult.fromExchange) {
       Log.debug(memo);
       memo.joinWithNewTrade(tradeResult);
-      memo.stopLossPrice = tradeResult.price * (1 - this.config.LossLimit);
+      memo.stopLimitPrice = tradeResult.price * (1 - this.config.StopLimit);
       this.store.setTrade(memo)
     }
     Log.alert(memo.tradeResult.toString())
   }
 
   private sellAndClose(memo: TradeMemo): void {
-    const tradeResult = this.exchange.marketSell(memo.tradeResult.symbol, memo.tradeResult.quantity);
-
+    const symbol = memo.tradeResult.symbol;
+    const tradeResult = this.exchange.marketSell(symbol, memo.tradeResult.quantity);
+    const gained = tradeResult.gained;
     if (tradeResult.fromExchange) {
       Log.debug(memo);
       const buyCommission = this.getBNBCommissionCost(memo.tradeResult.commission);
       const sellCommission = this.getBNBCommissionCost(tradeResult.commission);
       Log.info(`Commission: ~${buyCommission + sellCommission}`)
-      const profit = tradeResult.gained - memo.tradeResult.paid - sellCommission - buyCommission;
+      const profit = gained - memo.tradeResult.paid - sellCommission - buyCommission;
       tradeResult.profit = +profit.toFixed(2);
       memo.tradeResult = tradeResult;
       memo.setState(TradeState.SOLD)
@@ -133,24 +142,31 @@ export class V2Trader {
     } else {
       memo.hodl = true;
       memo.setState(TradeState.BOUGHT);
-      Log.alert(`An issue happened while selling ${memo.tradeResult.symbol}. The asset is marked HODL. Please, resolve it manually.`)
+      Log.alert(`An issue happened while selling ${symbol}. The asset is marked HODL. Please, resolve it manually.`)
     }
 
     this.store.setTrade(memo)
     Log.alert(tradeResult.toString());
-  }
 
-  private priceGoesUp(prices: PriceMemo, lastN: number = 3): boolean {
-    const lastPrices = prices.slice(-lastN);
-    if (lastPrices[0] == 0 || lastPrices.length < lastN) {
-      return false
+    if (memo.stateIs(TradeState.SOLD) && this.config.AveragingDown) {
+      this.afterAll.push(() => {
+        // all gains are reinvested to most unprofitable asset
+        // find a trade with the lowest profit percentage
+        const byProfitPercentDesc = (t1, t2) => t1.profitPercent() < t2.profitPercent() ? -1 : 1;
+        const lowestProfitTrade = this.store.getTradesList()
+          .filter(t => t.stateIs(TradeState.BOUGHT))
+          .sort(byProfitPercentDesc)[0];
+        if (lowestProfitTrade) {
+          Log.alert('Averaging down is enabled')
+          Log.alert(`All gains from selling ${symbol} are being invested to ${lowestProfitTrade.tradeResult.symbol}`);
+          this.buy(lowestProfitTrade, gained);
+        }
+      })
     }
-    return lastPrices.every((p, i) => i == 0 ? true : p > lastPrices[i - 1])
   }
 
   private getBNBCommissionCost(commission: number): number {
     const bnbPrice = this.prices["BNB" + this.config.PriceAsset];
     return bnbPrice ? commission * bnbPrice : 0;
   }
-
 }