+++ apps-script/Watcher.ts
@@ -41,6 +41,12 @@ function Ticker() {
     }
   })
 
+  try {
+    trader.afterTickerCheck();
+  } catch (e) {
+    Log.error(e)
+  }
+
   store.dumpChanges();
 
   Log.ifUsefulDumpAsEmail()